+++ src/lib/server/api/controllers/iam.controller.ts
@@ -8,13 +8,10 @@ import { signInEmailDto } from '../../../dtos/signin-email.dto';
 import { updateEmailDto } from '../../../dtos/update-email.dto';
 import { verifyEmailDto } from '../../../dtos/verify-email.dto';
 import { registerEmailDto } from '../../../dtos/register-email.dto';
-import { EmailVerificationsService } from '../services/email-verifications.service';
-import { LoginRequestsService } from '../services/login-requests.service';
 import type { HonoTypes } from '../common/types/hono.type';
 import type { Controller } from '../common/inferfaces/controller.interface';
 import { limiter } from '../middlewares/rate-limiter.middlware';
 import { requireAuth } from '../middlewares/auth.middleware';
-import TestJob from '../jobs/test.job';
 
 @injectable()
 export class IamController implements Controller {
@@ -22,29 +19,23 @@ export class IamController implements Controller {
 
 	constructor(
 		@inject(IamService) private iamService: IamService,
-		@inject(LoginRequestsService) private loginRequestsService: LoginRequestsService,
-		@inject(EmailVerificationsService) private emailVerificationsService: EmailVerificationsService,
 		@inject(LuciaProvider) private lucia: LuciaProvider,
-		@inject(TestJob) private testJob: TestJob
 	) { }
 
 	routes() {
 		return this.controller
 			.get('/user', async (c) => {
 				const user = c.var.user;
-				console.log('uwu')
-				this.testJob.queue('green');
-				// this.testJob.worker();
 				return c.json({ user: user });
 			})
 			.post('/login/request', zValidator('json', registerEmailDto), limiter({ limit: 10, minutes: 60 }), async (c) => {
 				const { email } = c.req.valid('json');
-				await this.loginRequestsService.create({ email });
+				await this.iamService.createLoginRequest({ email });
 				return c.json({ message: 'Verification email sent' });
 			})
 			.post('/login/verify', zValidator('json', signInEmailDto), limiter({ limit: 10, minutes: 60 }), async (c) => {
 				const { email, token } = c.req.valid('json');
-				const session = await this.loginRequestsService.verify({ email, token });
+				const session = await this.iamService.verifyLoginRequest({ email, token });
 				const sessionCookie = this.lucia.createSessionCookie(session.id);
 				setCookie(c, sessionCookie.name, sessionCookie.value, {
 					path: sessionCookie.attributes.path,
@@ -74,14 +65,14 @@ export class IamController implements Controller {
 			})
 			.patch('/email', requireAuth, zValidator('json', updateEmailDto), limiter({ limit: 10, minutes: 60 }), async (c) => {
 				const json = c.req.valid('json');
-				await this.emailVerificationsService.dispatchEmailVerificationRequest(c.var.user.id, json.email);
+				await this.iamService.dispatchEmailVerificationRequest(c.var.user.id, json.email);
 				return c.json({ message: 'Verification email sent' });
 			})
 			// this could also be named to use custom methods, aka /email#verify
 			// https://cloud.google.com/apis/design/custom_methods
 			.post('/email/verification', requireAuth, zValidator('json', verifyEmailDto), limiter({ limit: 10, minutes: 60 }), async (c) => {
 				const json = c.req.valid('json');
-				await this.emailVerificationsService.processEmailVerificationRequest(c.var.user.id, json.token);
+				await this.iamService.processEmailVerificationRequest(c.var.user.id, json.token);
 				return c.json({ message: 'Verified and updated' });
 			});
 	}
